Update man.bash
The command vi -c q ./readme.md opens the readme.md file and prompts the user to exit the editor. However, this command may save the file even if the user interrupts the editing process. This poses a security threat for files that may contain sensitive information. It is recommended to remove this line.